// types/SignUpInputType.d.ts
export default interface SignUpInputType {
  username: string;
  email: string;
  password: string;
  confirmPassword: string;
  firstName: string;
  lastName: string;
  age: string;
}

// export const initialSignUpInput = {
//   username: '',
//   email: '',
//   password: '',
//   confirmPassword: '',
//   firstName: '',
//   lastName: '',
//   age: '',
// };